+++ src/Model/CommandeDAO-sans-payum.php
@@ -5,21 +5,14 @@
 namespace Model;
 
 
-use Symfony\Component\Security\Core\User\UserInterface;
-use Symfony\Component\Security\Core\User\UserProviderInterface;
-use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
-use Symfony\Component\Security\Core\Exception\UnsupportedUserException;
-
 use Entity\Membre;
 use Entity\Cv;
 
 
-
 use Entity\Commande;
-use Payum\Core\Storage\StorageInterface;
-use Payum\Core\Model\Identity;
 
-class CommandeDAO extends DAO implements StorageInterface
+
+class CommandeDAO extends DAO
 {
 
     public function findAll(){
@@ -99,79 +92,6 @@ class CommandeDAO extends DAO implements StorageInterface
 
     }
 
-    /**
-     * devra retourner une nouvelle entité de paiement : une instance de
-     * Entity\Command
-     * @return object
-     */
-    public function create(){
-    $command = new \Entity\Command();
-    return $command;
-    }
-
-    /**
-     * @param object $model
-     *
-     * @return boolean
-     */
-    public function support($model){
-    $class = get_class($model);
-    return $class === '\Entity\Command';
-    }
-
-    /**
-     * @param object $model
-     *
-     * @throws \Payum\Core\Exception\InvalidArgumentException if not supported model given.
-     *
-     * @return void
-     */
-    public function update($model){
-    if( ! $this->support($model)) { 
-        throw \Payum\Core\Exception\InvalidArgumentException; 
-        }
-        $this->getDb()->query('UPDATE commande SET commande (...) ');
-    }
-
-    /**
-     * @param object $model
-     *
-     * @throws \Payum\Core\Exception\InvalidArgumentException if not supported model given.
-     *
-     * @return void
-     */
-    public function delete($model){
-        $this->getDb()->query('DELETE FROM commande WHERE commande = :commande');
-    }
-
-    /**
-     * @param mixed|IdentityInterface $id
-     *
-     * @return object|null
-     */
-    public function find($id){
-
-    }
-
-    /**
-     * @param array $criteria
-     *
-     * @return object[]
-     */
-    public function findBy(array $criteria){
-
-    }
-
-    /**
-     * @param object $model
-     *
-     * @throws \Payum\Core\Exception\InvalidArgumentException if not supported model given.
-     *
-     * @return IdentityInterface
-     */
-    public function identify($model){
-    return new Identity($model->getId(), get_class($model)); 
-    }
 }
 
  ?>
